feat(database): create `categories` related migration and model

+++ database/migrations/2024_04_25_063503_create_posts_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
   {
     Schema::create('posts', function (Blueprint $table) {
       $table->id('ID');
-      $table->integer('post_author')->unsigned()->default(0);
+      $table->bigInteger('post_author')->unsigned()->default(0);
       $table->text('post_title');
       $table->string('post_slug', 200)->unique();
       $table->longText('post_content');
